訪視紀錄(migration,seeder)

+++ database/migrations/2017_08_14_044716_create_interview_stu_questions_table.php
@@ -17,7 +17,7 @@ class CreateInterviewStuQuestionsTable extends Migration
 
             $table->increments('insQId');//AI
             $table->longText('insQuestion');//題目名稱
-            $table->tinyInteger('insQuestionVer');//題目版本
+            $table->integer('insQuestionVer');//題目版本
             $table->timestamps();
         });
     }
